Panic if CONFIG lock is already poisoned
Original issue is that anyhow errors require Send and Sync, but the
RwLock returns a Result<Guard, PoisonError<Guard>> and the guard is
`!Send`. This makes sense -- you can't claim the lock on one thread,
then send the lock guard to another thread to release the lock. So, the
PoisonError can't be made into an anyhow Error. However, if we get a
PoisonError, then another thread panicked while holding the lock, which
means the program is in a bad state and this thread should just panic
also.