Disallow absolute filepaths
We're preserving folder structure in cloud storage, so uploading
`dir/file` will create a file in a different location than doing `cd
dir` then uploading `file`. This may be confusing, particularly if we
allow appending files to a dataset after initial dataset creation,
because there's an unspoken assumption that subsequent bolster-upload
commands will be issued from the same directory that was used when
initially creating the dataset.